(refactor features): fix cart, wishlist and sprate public, private for anonymous. but still not done yet

+++ spring-microservices/course-service/src/main/java/com/hoangphihiep/dto/response/CartResponse.java
@@ -25,8 +25,8 @@ public class CartResponse {
         Integer courseId;
         String courseName;
         String authorName;
-        float rating;
-        float duration;
+        double rating;
+        double duration;
         String originalPrice;
         String currentPrice;
     }
